@@ -1,14 +1,14 @@
 #ifndef ZERO_BASE64_H
 #define ZERO_BASE64_H
 
+#include <span>
 #include <string>
 #include <vector>
 #include <optional>
 #include <cstddef>
-#include <nonstd/span.hpp>
 
 namespace zero::encoding::base64 {
-    std::string encode(nonstd::span<const std::byte> buffer);
+    std::string encode(std::span<const std::byte> buffer);
     std::optional<std::vector<std::byte>> decode(std::string_view encoded);
 }
 
