@@ -1,12 +1,12 @@
 #ifndef ZERO_NET_H
 #define ZERO_NET_H
 
+#include <span>
 #include <array>
 #include <string>
 #include <vector>
 #include <variant>
 #include <optional>
-#include <nonstd/span.hpp>
 
 namespace zero::os::net {
     struct IPv4Address {
@@ -24,8 +24,8 @@ namespace zero::os::net {
         std::vector<std::variant<IPv4Address, IPv6Address>> addresses;
     };
 
-    std::string stringify(nonstd::span<const std::byte, 4> ip);
-    std::string stringify(nonstd::span<const std::byte, 16> ip);
+    std::string stringify(std::span<const std::byte, 4> ip);
+    std::string stringify(std::span<const std::byte, 16> ip);
 
     std::optional<std::vector<Interface>> interfaces();
 }
